refactor auth thunks and added errorHandler + refreshTokenHandler
